phenoseries is a single script
