--- phenoseries/score.py
@@ -1,6 +1,0 @@
-# For now use this command
-# phenopy score phenoseries.mim_hpo_ids.txt --self --threads 48 >phenoseries.hrss.bma.pairwise.txt
-# phenopy score phenoseries.mim_hpo_ids.txt --self --scoring-method Resnik --threads 48 >phenoseries.resnik.bma.pairwise.txt
-# phenopy score phenoseries.mim_hpo_ids.txt --self --scoring-method Jaccard --threads 48 >phenoseries.jaccard.pairwise.txt
-
-# still need to add w2v scoring methods
